+++ src/main/php/scriptlet/HttpScriptlet.class.php
@@ -43,7 +43,7 @@ class HttpScriptlet extends \lang\Object {
   protected function _request() {
     return new HttpScriptletRequest();
   }
-  
+
   /**
    * Create a session object. Override this method to define
    * your own session object
@@ -73,7 +73,13 @@ class HttpScriptlet extends \lang\Object {
   protected function _url($url) {
     return new HttpScriptletURL($url);
   }
-  
+
+  /** @return scriptlet.Request */
+  public function request() { return $this->_request(); }
+
+  /** @return scriptlet.Response */
+  public function response() { return $this->_response(); }
+
   /**
    * Get authenticator for a certain request. Returns NULL in this default
    * implementation to indicate no authentication is required.
@@ -300,18 +306,6 @@ class HttpScriptlet extends \lang\Object {
   public function finalize() { }
   
   /**
-   * Set the request from the environment.
-   *
-   * @deprecated  Uses raw environment
-   * @param   scriptlet.HttpRequest request
-   */
-  protected function _setupRequest($request) {
-    $request->method= $request->getEnvValue('REQUEST_METHOD');
-    $request->setHeaders(getallheaders());
-    $request->setParams(array_merge($_GET, $_POST));
-  }    
-  
-  /**
    * This method is called to process any request and dispatches
    * it to on of the do* -methods of the scriptlet. It will also
    * call the `doCreateSession()` method if necessary.
@@ -426,25 +420,5 @@ class HttpScriptlet extends \lang\Object {
         $e
       );
     }
-  }
-
-  /**
-   * This method is called to process any request and dispatches
-   * it to on of the do* -methods of the scriptlet. It will also
-   * call the `doCreateSession()` method if necessary.
-   *
-   * @return  scriptlet.HttpScriptletResponse the response object
-   * @throws  scriptlet.ScriptletException indicating fatal errors
-   */
-  public function process() {
-    $request= $this->_request();
-    $response= $this->_response();
-
-    // Call service()
-    $this->_setupRequest($request);
-    $this->service($request, $response);
-    
-    // Return it
-    return $response;
   }
 }