+++ src/main/php/xp/scriptlet/WebConfiguration.class.php
@@ -2,6 +2,7 @@
 
 use util\Properties;
 use util\Hashmap;
+use lang\IllegalStateException;
 
 /**
  * Web application configuration
@@ -9,7 +10,7 @@ use util\Hashmap;
  * @see   xp://xp.scriptlet.WebApplication
  * @test  xp://scriptlet.unittest.WebConfigurationTest
  */
-class WebConfiguration extends \lang\Object {
+class WebConfiguration extends \lang\Object implements WebLayout {
   protected $prop= null;
   
   /**
@@ -84,7 +85,7 @@ class WebConfiguration extends \lang\Object {
   protected function configuredApp($profile, $application, $url) {
     $section= 'app::'.$application;
     if (!$this->prop->hasSection($section)) {
-      throw new \lang\IllegalStateException('Web misconfigured: Section '.$section.' mapped by '.$url.' missing');
+      throw new IllegalStateException('Web misconfigured: Section '.$section.' mapped by '.$url.' missing');
     }
 
     $app= new WebApplication($application);
@@ -95,20 +96,20 @@ class WebConfiguration extends \lang\Object {
 
     // Determine debug level
     $flags= WebDebug::NONE;
-    foreach ($this->readArray($profile, $section, 'debug', array()) as $lvl) {
+    foreach ($this->readArray($profile, $section, 'debug', []) as $lvl) {
       $flags |= WebDebug::flagNamed($lvl);
     }
     $app->setDebug($flags);
     
     // Initialization arguments
-    $app->setArguments($this->readArray($profile, $section, 'init-params', array()));
+    $app->setArguments($this->readArray($profile, $section, 'init-params', []));
  
     // Environment
     $app->setEnvironment($this->readHash($profile, $section, 'init-envs', new Hashmap())->toArray());
    
     return $app;
   }
-  
+
   /**
    * Gets all mapped applications
    *
@@ -118,7 +119,7 @@ class WebConfiguration extends \lang\Object {
    */
   public function mappedApplications($profile= null) {
     $mappings= $this->prop->readHash('app', 'mappings', null);
-    $apps= array();
+    $apps= [];
 
     // Verify configuration
     if (null === $mappings) {
@@ -133,7 +134,7 @@ class WebConfiguration extends \lang\Object {
     }
 
     if (0 === sizeof($apps)) {
-      throw new \lang\IllegalStateException('Web misconfigured: "app" section missing or broken');
+      throw new IllegalStateException('Web misconfigured: "app" section missing or broken');
     }
 
     return $apps;